Got the GET "/bugs" endpoint working with the app so that now the returned bugs are dynamically rendered using componentDidMount in the Dashboard.
I think this is the time where I should begin to incorporate Redux to
handle the state of the app, but not to contain any changes made to the
bugs' states. The bug states should be directly updated in the database
through the API, but the state of the app will be updated through an new
fetch request each time you go to the dashboard page.